School Application/requirements.txt
pygame==2.5.2
pandas==2.1.1
Pillow==10.0.0 